feat: исправлено удаление чужой карточки

+++ controllers/cards.js
@@ -1,5 +1,5 @@
 const {
-  ERROR_CODE_400, ERROR_CODE_404, ERROR_CODE_500, STATUS_201,
+  ERROR_CODE_404, ERROR_CODE_500, STATUS_201, ERROR_CODE_403,
 } = require('../utils/constants');
 const Card = require('../models/card');
 
@@ -11,31 +11,26 @@ module.exports.getCard = (req, res) => {
 };
 
 module.exports.deleteCardsById = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (card == null) {
         return res.status(ERROR_CODE_404).send({ message: 'Передан несуществующий _id карточки.' });
       }
-      return res.send({ data: card });
-    })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(ERROR_CODE_400).send({ message: 'Некоректные данные _id карточки.' });
+      if (!(card.owner._id.toString() === req.user._id)) {
+        return res.status(ERROR_CODE_403).send({ message: 'Вы не можете удалить чужую карточку!' });
       }
-      return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка' });
-    });
+      return Card.findByIdAndRemove(req.params.cardId)
+        .then((cardDelete) => res.send({ data: cardDelete }))
+        .catch(() => res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка' }));
+    })
+    .catch(() => res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка' }));
 };
 
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
   Card.create({ name, owner: req.user._id, link })
     .then((card) => res.status(STATUS_201).send({ data: card }))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res.status(ERROR_CODE_400).send({ message: 'Переданы некорректные данные при создании карточки.' });
-      }
-      return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка' });
-    });
+    .catch(() => res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка' }));
 };
 
 module.exports.likeCard = (req, res) => Card.findByIdAndUpdate(
@@ -52,12 +47,7 @@ module.exports.likeCard = (req, res) => Card.findByIdAndUpdate(
     }
     return res.send({ data: card });
   })
-  .catch((err) => {
-    if (err.name === 'CastError') {
-      return res.status(ERROR_CODE_400).send({ message: 'Переданы некорректные данные для постановки лайка.' });
-    }
-    return res.status(ERROR_CODE_500).send({ message: `Произошла ошибка ${err.status}` });
-  });
+  .catch((err) => res.status(ERROR_CODE_500).send({ message: `Произошла ошибка ${err.status}` }));
 
 module.exports.dislikeCard = (req, res) => Card.findByIdAndUpdate(
   req.params.cardId,
@@ -73,9 +63,4 @@ module.exports.dislikeCard = (req, res) => Card.findByIdAndUpdate(
     }
     return res.send({ data: card });
   })
-  .catch((err) => {
-    if (err.name === 'CastError') {
-      return res.status(ERROR_CODE_400).send({ message: 'Переданы некорректные данные для снятии лайка. ' });
-    }
-    return res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка' });
-  });
+  .catch(() => res.status(ERROR_CODE_500).send({ message: 'Произошла ошибка' }));
